IB/core: Expand nvme_queue_size field according to spec
According to NVMe specification, the maximal NVMe device queue
size is 2^16 (17 bits), so expand the field accordingly.

+++ include/rdma/ib_verbs.h
@@ -1062,7 +1062,7 @@ struct ib_nvmf_init_data {
 	u8				staging_buffer_log_page_size;
 	u16				staging_buffer_number_of_pages;
 	u32				staging_buffer_page_offset;
-	u16				nvme_queue_size;
+	u32				nvme_queue_size;
 	u64				*staging_buffer_pas;
 };
